Moved XLSX row generator to a utility file
